requirements.txt
# Dépendances principales pour le projet de scoring de crédit

# Data Science et ML
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
shap>=0.41.0

# Visualisation
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# API et Web
fastapi>=0.95.0
uvicorn>=0.20.0
streamlit>=1.25.0

# Configuration et utilitaires
pyyaml>=6.0
pyarrow>=10.0.0
pydantic>=1.10.0

# Tests
pytest>=7.0.0
pytest-cov>=4.0.0

# Formatage et qualité du code
black>=22.0.0
flake8>=5.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Autres
scipy>=1.9.0
statsmodels>=0.13.0



